requirements.txt
Flask
matplotlib
gunicorn
supabase-py